pyproject.toml
[tool.poetry]
name = "nrdtech-python-aws-lambda"
version = "0.1.0"
description = ""
authors = ["Nic Delorme <nic.delorme@nrdtech.io>"]
license = "Apache License 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.88"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
mangum = "^0.19.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^8.3.4"
python-dotenv = "^1.0.1"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
